Added the Ability to Use Decimals and Improved Clear Entry
Decimals can now be used in the user's math problems. Also, the clear entry button was refactored with the inclusion of the decimal. I'm going to work on
adding the ability to use keyboard inputs as well. This will improve the user's experience with using the calculator, making it easier and faster to perform
calculations.